+++ crates/c-api/Cargo.toml
@@ -48,3 +48,4 @@ disable-logging = ["log/max_level_off", "tracing/max_level_off"]
 coredump = ["wasmtime/coredump"]
 addr2line = ["wasmtime/addr2line"]
 demangle = ["wasmtime/demangle"]
+threads = ["wasmtime/threads"]
