+++ crates/cli-flags/Cargo.toml
@@ -28,3 +28,5 @@ parallel-compilation = ["wasmtime/parallel-compilation", "dep:rayon"]
 logging = ["dep:file-per-thread-logger", "dep:tracing-subscriber"]
 cranelift = ["wasmtime/cranelift"]
 coredump = ["wasmtime/coredump"]
+gc = ["wasmtime/gc"]
+threads = ["wasmtime/threads"]
