+++ crates/cranelift/Cargo.toml
@@ -38,3 +38,4 @@ component-model = ["wasmtime-environ/component-model"]
 incremental-cache = ["cranelift-codegen/incremental-cache"]
 wmemcheck = []
 gc = ["wasmtime-environ/gc"]
+threads = ["wasmtime-environ/threads"]
