+++ crates/runtime/Cargo.toml
@@ -72,3 +72,4 @@ wmemcheck = ['dep:wasmtime-wmemcheck']
 debug-builtins = ['dep:wasmtime-jit-debug']
 gc = ["wasmtime-environ/gc"]
 coredump = ["dep:wasm-encoder"]
+threads = ["wasmtime-environ/threads"]
