+++ crates/runtime/src/lib.rs
@@ -20,13 +20,15 @@ mod instance;
 mod memory;
 mod mmap;
 mod mmap_vec;
-mod parking_spot;
 mod send_sync_ptr;
 mod store_box;
 mod sys;
 mod table;
 mod traphandlers;
 mod vmcontext;
+
+mod threads;
+pub use self::threads::*;
 
 #[cfg(feature = "debug-builtins")]
 pub mod debug_builtins;
@@ -48,9 +50,7 @@ pub use crate::instance::{
 pub use crate::instance::{
     InstanceLimits, PoolingInstanceAllocator, PoolingInstanceAllocatorConfig,
 };
-pub use crate::memory::{
-    DefaultMemoryCreator, Memory, RuntimeLinearMemory, RuntimeMemoryCreator, SharedMemory,
-};
+pub use crate::memory::{DefaultMemoryCreator, Memory, RuntimeLinearMemory, RuntimeMemoryCreator};
 pub use crate::mmap::Mmap;
 pub use crate::mmap_vec::MmapVec;
 pub use crate::mpk::MpkEnabled;
