+++ crates/wasi-threads/Cargo.toml
@@ -19,4 +19,4 @@ anyhow = { workspace = true }
 log = { workspace = true }
 rand = "0.8"
 wasi-common = { workspace = true, features = ["exit"]}
-wasmtime = { workspace = true }
+wasmtime = { workspace = true, features = ['threads'] }
